+++ internal/daemon/worker/controller_connection.go
@@ -434,13 +434,13 @@ func monitorUpstreamConnectionState(ctx context.Context, cc *grpc.ClientConn, co
 	}
 
 	for cc.WaitForStateChange(ctx, state) {
-		newState := cc.GetState()
+		state = cc.GetState()
 
 		// if the client is shutdown, exit function
-		if newState == connectivity.Shutdown {
+		if state == connectivity.Shutdown {
 			return
 		}
 
-		connectionState.Store(newState)
+		connectionState.Store(state)
 	}
 }